requirements.txt
# Project149 Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
bcrypt==4.1.1

# Auth
python-jose[cryptography]==3.3.0
pyjwt==2.8.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# ML/Recommendations
scikit-learn==1.3.2

# AI/Search
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: Redis
redis==5.0.1
